+++ scripts/optimize.py
@@ -59,20 +59,19 @@ def generate_items(n):
 
 
 # Worker
-def test(items):
+def test(items, chunk_size):
     t1 = time.monotonic()
-    ids = account.calendar.bulk_create(items=items)
+    ids = account.calendar.bulk_create(items=items, chunk_size=chunk_size)
     t2 = time.monotonic()
-    account.bulk_delete(ids)
+    account.bulk_delete(ids=ids, chunk_size=chunk_size)
     t3 = time.monotonic()
 
     delta1 = t2 - t1
     rate1 = len(ids) / delta1
     delta2 = t3 - t2
     rate2 = len(ids) / delta2
-    print(('Time to process %s items (batchsize %s/%s, poolsize %s): %s / %s (%s / %s per sec)' % (
-        len(ids), services.CreateItem.CHUNKSIZE, services.DeleteItem.CHUNKSIZE,
-        config.protocol.poolsize, delta1, delta2, rate1, rate2)))
+    print(('Time to process %s items (batchsize %s, poolsize %s): %s / %s (%s / %s per sec)' % (
+        len(ids), chunk_size, config.protocol.poolsize, delta1, delta2, rate1, rate2)))
 
 
 # Generate items
@@ -80,16 +79,14 @@ calitems = list(generate_items(500))
 
 print('\nTesting batch size')
 for i in range(1, 11):
-    services.CreateItem.CHUNKSIZE = 25 * i
-    services.DeleteItem.CHUNKSIZE = 25 * i
+    chunk_size = 25 * i
     config.protocol.poolsize = 5
-    test(calitems)
+    test(calitems, chunk_size)
     time.sleep(60)  # Sleep 1 minute. Performance will deteriorate over time if we give the server tie to recover
 
 print('\nTesting pool size')
 for i in range(1, 11):
-    services.CreateItem.CHUNKSIZE = 10
-    services.DeleteItem.CHUNKSIZE = 10
+    chunk_size = 10
     config.protocol.poolsize = i
-    test(calitems)
+    test(calitems, chunk_size)
     time.sleep(60)