+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ HEAD
 * Added ``Message.reply()``, ``Message.reply_all()``, and ``Message.forward()`` methods.
 * The full search API now works on single folders *and* collections of folders, e.g.
   ``some_folder.glob('foo*').filter()``, ``some_folder.children.filter()`` and ``some_folder.walk().filter()``.
+* Deprecated ``EWSService.CHUNKSIZE`` in favor of a per-request chunk_size available on ``Account.bulk_foo()`` methods.
 
 
 1.10.7